refactor: change newhousehold from comp to view
