+++ frontend/src/views/registered/NewHousehold.vue
